+++ CHANGELOG.md
@@ -33,10 +33,10 @@
 
 ## 3.4.1
 ### Added
-
 - Added `obj` setter to context builder as an alternative to `currentContext.obj`
 - Added `option().boolean()` and `argument().boolean()`
 - `uint()` and `ulong()` parameter type conversions.
+- `CliktCommand.test` extension for testing your commands and their output
 
 ### Changed
 - Updated Kotlin to 1.6.20